feat: take the whole family's new releases, and widen the timestamps with them (#77)
Twelve crates went out today. The bundles take all of them:

  am-fs-core 0.2.4 -> 0.2.5     am-fs-ext4 0.3.3 -> 0.5.1
  am-fs-ntfs 0.3.3 -> 0.4.0     am-fs-xfs 0.5.0 -> 0.7.0
  am-fs-btrfs 0.5.0 -> 0.6.2    am-fs-erofs 0.1.1 -> 0.1.5
  am-fs-squashfs 0.1.1 -> 0.1.5 am-partitions 0.4.0 -> 0.4.1
  am-img-qcow2 0.4.2 -> 0.4.5   am-img-{vhd,vhdx,vmdk} 0.3.2 -> 0.3.5

Most of that is hardening: values that came off an image were used to
allocate, to address or to bound a loop before being checked. The two
that change this app's own code are timestamps.

SECONDS ARE SIGNED AND SIXTY-FOUR BITS WIDE NOW, on both sides of the
call. ext4's on-disk timestamp is a signed 32-bit field that its
matching *_extra field extends by two more bits, so the real range is
roughly 1901 to 2446 -- and the driver widened `fs_ext4_attr_t` and
`fs_ext4_utimens` to match it.

Held as `UInt32` here, both ends were wrong. Reading, every date before
1970 came back as one in the far future and everything past 2038 was
truncated. Writing, `utimens` clamped -- and the top of that clamp,
`UInt32.max`, was the driver's "leave this one unchanged" sentinel, so
setting a far-future time silently set nothing at all. Neither showed
up as an error; both showed up as a wrong date in the Finder.

`BackendFileAttributes` carries `Int64` seconds, and `utimens` passes
`tv_sec` through untouched and names the sentinel rather than colliding
with it.

Every target builds: the six extensions, the library and the app.

+++ DiskJockeyEXT4/FileSystemBackend.swift
@@ -30,10 +30,19 @@ struct BackendFileAttributes {
     var gid: UInt32
     var size: UInt64
     var linkCount: UInt16
-    var atime: UInt32
-    var mtime: UInt32
-    var ctime: UInt32
-    var crtime: UInt32
+    /// SECONDS ARE SIGNED AND SIXTY-FOUR BITS WIDE, matching the driver
+    /// as of am-fs-ext4 0.5.0 and the on-disk field it comes from.
+    ///
+    /// ext4's base timestamp is a signed 32-bit value that its matching
+    /// `*_extra` field extends by two more bits, so the real range is
+    /// roughly 1901 to 2446. Held as `UInt32` here, everything before
+    /// 1970 read back as a date in the far future and everything after
+    /// 2038 was truncated -- and neither showed up as an error, only as
+    /// a wrong date in the Finder.
+    var atime: Int64
+    var mtime: Int64
+    var ctime: Int64
+    var crtime: Int64
 }
 
 struct BackendDirectoryEntry {
